Update CHANGELOG to reflect v0.1.6 binary release
Providing binary builds for the v0.1.6 release was mostly a last
minute decision. I've been wanting to do this for a while, but
hadn't yet pulled the trigger. I forced myself to do so with the
v0.1.6 release, but the decision was made *after* I had already
tagged the release. Updating the CHANGELOG so that the `master`
branch reflects the v0.1.6 release via GitHub.

- Explicitly note Go version used to generate binaries
- Fix section order to match existing pattern
- Note binary platforms