Add IP range cert scanner prototype: certsum
Initial prototype to scan one or more given IP ranges and
and report overview of certificate status. This is a "MVP"
build and considered to be of "alpha" level quality.
Many of the exposed flags, help text and summary output
is subject to change significantly in later releases.

Performance is so-so, with only the initial port scanning
handled in a concurrent fashion; the actual cert retrieval
process is handled one host/port at a time. Unsurprisingly,
this does not scale well for larger IP ranges.

Additionally, specifying partial/non-CIDR IP ranges is not
tested or supported at this time, though it is possible
to specify multiple CIDR ranges for evaluation.

This build also allows for filtering "OK" status hosts and
certs to either increase or reduce the amount of information
provided. Two summary modes are also provided in order to
control the level of detail in the provided summary.

Other changes include updating flag handling based on a
specified "app type", one of `Inspecter`, `Plugin` or
`Scanner` and refactoring of shared functionality to reduce
code duplication.

Lastly, a sweep of existing documenation has been made to fix
minor issues and provide clarification between existing
functionality and new. Much work remains to be done.

refs GH-103