zerolog logger: Add missing Msg() call
Add missing Msg() invocation for logger call associated with a
port scan error. Without this, the logger call fails to emit
the intended error message. Because the most common error case
for a port scan is an i/o timeout, we emit the error message at
debug level and continue scanning the next port.