Add annotation to context deadline error message
Wrap this specific error message with additional details to
make clear that the error is due to the plugin runtime exceeding
the specified (or default) timeout value.

As a side-effect of adding this support, this commit also
provides a baseline Nagios Plugin Performance Data metric that
tracks plugin execution time.

refs GH-285