Fix go install steps to include valid soruce path
The prior directions assumed that check_cert would be placed
in the current working directory via go install. Update the
directions to explicitly set `GOBIN` so that the generated
binary will be placed in the current working directory as
intended.

Reference the output from `go help install` to explain why
we are setting GOBIN explicitly.