Add certificate perfdata metrics
Add performance data metrics to track upcoming expirations
and total certificate types (scopes) in the chain:

- `expires_leaf`
- `expires_intermediate`
- `certs_present_leaf`
- `certs_present_intermediate`
- `certs_present_root`
- `certs_present_unknown`

As part of providing these performance data metrics various helper
functions were added:

- `certs.NumLeafCerts`
- `certs.NumIntermediateCerts`
- `certs.NumRootCerts`
- `certs.NumUnknownCerts`
- `certs.LeafCerts`
- `certs.IntermediateCerts`
- `certs.RootCerts`
- `certs.OldestLeafCert`
- `certs.OldestIntermediateCert`
- `certs.OldestRootCert`
- `certs.ExpiresInDays`

The README file has been updated to list the purpose of each
performance data metric.

As of this commit, failure to collect performance data is emitted as
an error message and recorded in the plugin's error collection (for
display in `LongServiceOutput`). Future work is scheduled to revisit
this choice and audit exit states as a whole to determine if more
appropriate exit states should be used.

refs GH-445
refs GH-464