Makefile: Replace most recursively expanded vars
Based on some light reading, inconsistent use of variable types is
causing not only a performance hit in some cases, but likely some
unreliable results based on how the recursive expanded variable type
is resolved.

For any case where the values are fixed or should be fixed after an
initial interpolation those variables should be updated to the simply
expanded type.

This commit reflects those changes.

refs GH-476