Update versioned build image source syntax
The dependabot/docker/builds/Dockerfile was set to use
this image:

- go-ci-oldstable-build-v0.8.0-2-g87114607

and earlier an updated image was released, but Dependabot
refused to recognize it as an update:

- go-ci-oldstable-build-v0.8.1-0-ga96dbbdb

I've released alternate tags without the "long" git tag
details closer to what Dependabot parses for other
Dockerfiles that it monitors.

Updating the Dockerfile to that simpler version tag as
a troubleshooting step.