Add TODO to open upstream RFE re unexpected error
The `ErrX509CertReliesOnCommonName` sentinel error is used
by this project to aid in detecting this specific error
scenario. While performing a string match for this error
works at present, it feels brittle and subject to break
in the future if the error text should change.

By opening a RFE/proposal in the upstream Go project, we may
gain a supported/stable approach for matching the error in
the future.

refs GH-520