Adopt isResultIgnored for expiration validation
Use this small helper function to mirror the recent change
to the certs.ValidateHostname function. Using this helper
function is intended to clarify the logic regarding whether
the validation result will be ignored.

It mostly ensures that the validation logic is consistent for
each point where a validation result is returned.

refs GH-505