Add dependency on libselinux-utils package
Update nfpm configurations to set explicit dependency on
the libselinux-utils package for generated RPM packages.

This ensures that the SELinux related commands used by
post-installation scripts are available.

NOTE: The `libselinux-utils` package is installed by default
when using the `minimal` installation option, so this is
mostly "just to be sure".

refs GH-509