Implement certScanner rate limiting
This function (used as a goroutine) previously received the rate
limiter channel intended for this purpose, but failed to make use of
it.

The revive linter upgraded in the golangci-lint v1.52.0 release flagged
the channel as an unused parameter which surfaced this issue.

This commit implements rate limiting for this function similar to what
is already applied to the portScanner function.

Small tweaks have been made to debug logging for both functions to
help troubleshoot potential future issues with adding/removing rate
limiting reservations.

refs GH-552