Update `lscert` hostname validation behavior
Update hostname validation logic to ignore the result if
a leaf certificate is not present in the given certificate
chain OR if the `dns-name` flag was not used. Add a slight
addendum/note explaining what is needed (if a leaf cert is
present) or that the check is unsupported (if no leaf cert
is present).

Add new functions to the `internal/certs` package to quickly
answer whether any certificate of the specified type is
present within a given certificate chain:

- `HasLeafCert`
- `HasIntermediateCert`
- `HasRootCert`

refs GH-952