Refactor ExpirationValidationResult status logic
Store a filtered certificate chain value as part of generating
the ExpirationValidationResult value. Use that when providing
a filtered certificate chain.

Rework how the validation status is determined: use the stored
filtered certificate chain instead of bundling separate logic
to fallback to the original cert chain if one is not provided
at time of method call.

Minor cleanup of Status and StatusDetail methods.