Return ErrNoCertsFound error when chain is empty
This alters a recent change in behavior to return
the `ErrIncompleteCertificateChain` error in place of the
more generic `ErrMissingValue` error value previously used.

Now, we return the more appropriate `ErrNoCertsFound` error
value from all current validation checks when an empty
certificate chain is detected.

Some further work is needed to update WARNING and CRITICAL
state detection to look for an `ErrNoCertsFound` error and
treat it as an appropriate state.

- refs GH-1155
- refs GH-1167